Update back-end to download Elering data to database
New methods to obtain electircity data from Elering.
If data for that day is already in database it will skip Elering data API and just return the data for that day.
If the data is not in database it will save it to database and then return the valid data.